pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "airflow-powerbi-provider"
version = "1.0.5"
authors = [
  { name="Glenn Schuurman", email="info@schuurman-it.com" },
]
description = "A provider package for Apache Airflow to refresh Powerbi Datasets"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "apache-airflow>=2.9.3",
    "azure-identity>=1.17.1"
]

[tool.hatch.build.targets.wheel]
packages = ["src/PowerBI_Operator"]

[project.urls]
Homepage = "https://github.com/gschuurman/airflow-powerbi-provider"
Issues = "https://github.com/gschuurman/airflow-powerbi-provider/issues"
